Moved user data from summary to the external component
